+++ src/entities/eyeDroids/eyeDroid.h
@@ -23,13 +23,13 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 extern void dropCarriedItem(void);
 extern int getDistance(int x1, int y1, int x2, int y2);
 extern double randF(void);
-extern void addRandomWeapon(float x, float y);
+extern void addRandomWeapon(int x, int y);
 extern float limit(float i, float a, float b);
 extern void playSound(int snd, int ch);
 extern void updateObjective(char *targetName);
 extern int enemyCanSeePlayer(Entity *e);
 extern void fireTriggers(char *name);
-extern void addRandomItems(float x, float y);
+extern void addRandomItems(int x, int y);
 extern int rrnd(int low, int high);
 extern void addExplosion(float x, float y, int radius, Entity *owner);
 extern void throwDebris(float x, float y, int amount);
